+++ src/index.ts
@@ -1,4 +1,4 @@
 export { default as ComponentParser } from "./ComponentParser";
 export { cli } from "./cli";
-export { default } from "./plugin";
+export { default, type SveldWarning } from "./plugin";
 export { sveld } from "./sveld";